Adapter Design Pattern Bank Example
Lloyds bank is an international bank offers services worldwide. For
offshore account holders, the tax rate is 0.03%. And, in India it offers
two types of accounts, Standard and Platinum. Tax rules are not applied
to indian bank accounts. Now the offshore bank is incompatible to Indian
account types. We need to design an AccountAdapter to make both the
incompatible account types to work together.

In the above example, the interaction diagram will be represented as
shown below. Here client is just calling the getBalance() method of
adapter. Adapter calls getOffshoreBalance() on adaptee and returns the
result that client is expecting. The getBalance() methods inside Adapter
will calculate the account balance by deducing the taxes applied for the
offshore account.